requirements.txt
pandas==2.2.1
pdfplumber==0.10.3
python-dateutil==2.8.2
python-dotenv==1.0.0
ipykernel>=6.29.3
transformers
sentencepiece
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
langchain
langchain-community
langchain-core
langchain-experimental
langchain-huggingface
faiss-cpu
pinecone
tensorflow==2.15.0
tf-keras==2.15.0
protobuf>=3.20.3,<4.21.0
notebook>=7.0.0
ipywidgets>=8.1.0
jupyterlab_widgets>=3.0.0
widgetsnbextension>=4.0.0
faiss-cpu